Renamed the `network_interface` and `net_namespace` modules
