+++ src/main/host/descriptor/socket/inet/udp.rs
@@ -16,7 +16,7 @@ use crate::host::descriptor::{
     SyscallResult,
 };
 use crate::host::memory_manager::MemoryManager;
-use crate::host::network::net_namespace::NetworkNamespace;
+use crate::host::network::namespace::NetworkNamespace;
 use crate::host::syscall::io::{write_partial, IoVec, IoVecReader};
 use crate::host::syscall_types::SyscallError;
 use crate::network::packet::{PacketRc, PacketStatus};
