+++ src/main/host/descriptor/socket/unix.rs
@@ -19,7 +19,7 @@ use crate::host::descriptor::{
     SyscallResult,
 };
 use crate::host::memory_manager::MemoryManager;
-use crate::host::network::net_namespace::NetworkNamespace;
+use crate::host::network::namespace::NetworkNamespace;
 use crate::host::syscall::io::{IoVec, IoVecReader, IoVecWriter};
 use crate::host::syscall_types::SyscallError;
 use crate::utility::callback_queue::{CallbackQueue, Handle};
