+++ src/main/host/host.rs
@@ -35,8 +35,8 @@ use crate::core::work::task::TaskRef;
 use crate::core::worker::Worker;
 use crate::cshadow;
 use crate::host::descriptor::socket::abstract_unix_ns::AbstractUnixNamespace;
-use crate::host::network::net_namespace::NetworkNamespace;
-use crate::host::network::network_interface::{NetworkInterface, PcapOptions};
+use crate::host::network::interface::{NetworkInterface, PcapOptions};
+use crate::host::network::namespace::NetworkNamespace;
 use crate::host::process::Process;
 use crate::host::thread::ThreadId;
 use crate::network::relay::{RateLimit, Relay};
